bhg/pyproject.toml
[project]
name = "bhg"
version = "0.1.0"
description = "Medical Chatbot Bias & Harm Evaluation"
requires-python = ">=3.11,<3.13"
dependencies = [
    "deepeval>=3.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "litellm>=1.0.0",
    "google-genai>=1.0.0",
]
